Ejemplos/RaspberryPi/Basico/GPIO_Basics/requirements.txt
# Control GPIO
RPi.GPIO==0.7.1

# Utilidades
schedule==1.2.0
python-dotenv==1.0.0 